LPS-138789 Rename com.liferay.remote.app.admin.web to com.liferay.remote.app.web
