+++ modules/apps/remote-app/remote-app-web/bnd.bnd
@@ -1,4 +1,4 @@
 Bundle-Name: Liferay Remote App Admin Web
-Bundle-SymbolicName: com.liferay.remote.app.admin.web
+Bundle-SymbolicName: com.liferay.remote.app.web
 Bundle-Version: 2.0.6
 Web-ContextPath: /remote-app-web
