+++ modules/apps/remote-app/remote-app-web/src/main/java/com/liferay/remote/app/web/internal/application/list/RemoteAppAdminPanelApp.java
@@ -12,13 +12,13 @@
  * details.
  */
 
-package com.liferay.remote.app.admin.web.internal.application.list;
+package com.liferay.remote.app.web.internal.application.list;
 
 import com.liferay.application.list.BasePanelApp;
 import com.liferay.application.list.PanelApp;
 import com.liferay.application.list.constants.PanelCategoryKeys;
 import com.liferay.portal.kernel.model.Portlet;
-import com.liferay.remote.app.admin.web.internal.constants.RemoteAppAdminPortletKeys;
+import com.liferay.remote.app.web.internal.constants.RemoteAppAdminPortletKeys;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
