+++ modules/apps/remote-app/remote-app-web/src/main/java/com/liferay/remote/app/web/internal/frontend/taglib/clay/data/set/RemoteAppEntryClayDataSetActionProvider.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.remote.app.admin.web.internal.frontend.taglib.clay.data.set;
+package com.liferay.remote.app.web.internal.frontend.taglib.clay.data.set;
 
 import com.liferay.frontend.taglib.clay.data.set.ClayDataSetActionProvider;
 import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItem;
@@ -29,7 +29,7 @@ import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
 import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.remote.app.admin.web.internal.constants.RemoteAppAdminConstants;
+import com.liferay.remote.app.web.internal.constants.RemoteAppAdminConstants;
 
 import java.util.List;
 import java.util.ResourceBundle;
