+++ modules/apps/remote-app/remote-app-web/src/main/java/com/liferay/remote/app/web/internal/frontend/taglib/clay/data/set/provider/RemoteAppEntryClayDataSetDataProvider.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.remote.app.admin.web.internal.frontend.taglib.clay.data.set.provider;
+package com.liferay.remote.app.web.internal.frontend.taglib.clay.data.set.provider;
 
 import com.liferay.frontend.taglib.clay.data.Filter;
 import com.liferay.frontend.taglib.clay.data.Pagination;
@@ -21,8 +21,8 @@ import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.search.Sort;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.remote.app.admin.web.internal.constants.RemoteAppAdminConstants;
-import com.liferay.remote.app.admin.web.internal.frontend.taglib.clay.data.set.RemoteAppClayDataSetEntry;
+import com.liferay.remote.app.web.internal.constants.RemoteAppAdminConstants;
+import com.liferay.remote.app.web.internal.frontend.taglib.clay.data.set.RemoteAppClayDataSetEntry;
 import com.liferay.remote.app.model.RemoteAppEntry;
 import com.liferay.remote.app.service.RemoteAppEntryLocalService;
