+++ modules/apps/remote-app/remote-app-web/src/main/java/com/liferay/remote/app/web/internal/portlet/action/EditRemoteAppEntryMVCRenderCommand.java
@@ -12,13 +12,13 @@
  * details.
  */
 
-package com.liferay.remote.app.admin.web.internal.portlet.action;
+package com.liferay.remote.app.web.internal.portlet.action;
 
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCRenderCommand;
 import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.remote.app.admin.web.internal.constants.RemoteAppAdminPortletKeys;
-import com.liferay.remote.app.admin.web.internal.constants.RemoteAppAdminWebKeys;
-import com.liferay.remote.app.admin.web.internal.display.context.RemoteAppAdminDisplayContext;
+import com.liferay.remote.app.web.internal.constants.RemoteAppAdminPortletKeys;
+import com.liferay.remote.app.web.internal.constants.RemoteAppAdminWebKeys;
+import com.liferay.remote.app.web.internal.display.context.RemoteAppAdminDisplayContext;
 import com.liferay.remote.app.service.RemoteAppEntryLocalService;
 
 import javax.portlet.PortletException;
