+++ modules/apps/remote-app/remote-app-web/src/main/resources/META-INF/resources/admin/init.jsp
@@ -20,7 +20,7 @@
 page import="com.liferay.portal.kernel.bean.BeanPropertiesUtil" %><%@
 page import="com.liferay.portal.kernel.language.LanguageUtil" %><%@
 page import="com.liferay.portal.kernel.util.ParamUtil" %><%@
-page import="com.liferay.remote.app.admin.web.internal.constants.RemoteAppAdminConstants" %><%@
-page import="com.liferay.remote.app.admin.web.internal.constants.RemoteAppAdminWebKeys" %><%@
-page import="com.liferay.remote.app.admin.web.internal.display.context.RemoteAppAdminDisplayContext" %><%@
+page import="com.liferay.remote.app.web.internal.constants.RemoteAppAdminConstants" %><%@
+page import="com.liferay.remote.app.web.internal.constants.RemoteAppAdminWebKeys" %><%@
+page import="com.liferay.remote.app.web.internal.display.context.RemoteAppAdminDisplayContext" %><%@
 page import="com.liferay.remote.app.exception.DuplicateRemoteAppEntryException" %>
